Time: 0 ms (100%), Space: 10.1 MB (11.55%) - LeetHub

+++ 0088-merge-sorted-array/0088-merge-sorted-array.c
@@ -78,7 +78,7 @@ void merge(int* nums1, int nums1Size, int m, int* nums2, int nums2Size, int n) {
     //[1,2,2,3,5,6]
     //int temp = nums1[0];;
     int* arr = (int*)malloc((m+n) * sizeof(int)); //  Valid in C
-    int arrIndex = 0;
+    short arrIndex = 0;
     //int a, b;
     //int push;
     short i = 0, j = 0;
